// src/lib/api/models/UserDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserDto = {
    id?: number;
    username?: string;
    email?: string;
    fullName?: string;
    imageUrl?: string;
    score?: number;
    subId?: string;
};